Organize resource list components into 2 sub families
