+++ src/Components/ResourceList/Card/ResourceCardName.scss
