+++ src/Components/ResourceList/List/ResourceList.scss
@@ -1,0 +1,87 @@
+@import "../../../Styles/modules";
+
+/*
+  Resource List
+  ------------------------------------------------------------------------------
+*/
+
+.resource-list {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+}
+
+/*
+  Header & Sorting
+  ------------------------------------------------------------------------------
+*/
+
+.resource-list--header {
+  display: flex;
+  align-items: center;
+  padding-bottom: $cf-marg-c;
+  justify-content: space-between;
+}
+
+.resource-list--sorting {
+  display: flex;
+  align-items: center;
+}
+
+.resource-list--filter {
+  min-width: 10px;
+}
+
+.resource-list--sorter {
+  user-select: none;
+  font-size: $form-md-font;
+  font-weight: 600;
+  text-transform: uppercase;
+  color: $g11-sidewalk;
+  transition: color 0.25s ease;
+  margin-left: $cf-marg-c;
+  display: flex;
+  align-items: center;
+  align-content: center;
+
+  &:hover {
+    color: $g18-cloud;
+    cursor: pointer;
+  }
+
+  &.resource-list--sort-descending,
+  &.resource-list--sort-ascending {
+    color: $c-pool;
+  }
+}
+
+.resource-list--sort-arrow {
+  width: $cf-marg-c;
+  height: $cf-marg-c;
+  position: relative;
+  margin-left: $cf-marg-a;
+  transition: opacity 0.25s ease;
+  opacity: 0;
+
+  .resource-list--sort-descending &,
+  .resource-list--sort-ascending & {
+    opacity: 1;
+  }
+
+  > span.icon {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    transition: transform 0.25s ease;
+  }
+
+  .resource-list--sort-descending & > span.icon {
+    transform: translate(-50%, -50%) rotate(0deg);
+  }
+
+  .resource-list--sort-ascending & > span.icon {
+    transform: translate(-50%, -50%) rotate(180deg);
+  }
+}
