+++ src/Components/ResourceList/List/ResourceListBody.tsx
@@ -3,7 +3,7 @@ import React, {PureComponent, ReactNode} from 'react'
 import classnames from 'classnames'
 
 // Types
-import {StandardProps} from '../../Types'
+import {StandardProps} from '../../../Types'
 
 interface Props extends StandardProps {
   /** Element to show when no children are passed in, useful for implementing filtering */
