+++ src/index.ts
@@ -71,12 +71,12 @@ export * from './Components/IndexList/IndexListHeader'
 export * from './Components/IndexList/IndexListHeaderCell'
 export * from './Components/IndexList/IndexListRow'
 export * from './Components/IndexList/IndexListRowCell'
-export * from './Components/ResourceList/ResourceList'
-export * from './Components/ResourceList/ResourceListBody'
-export * from './Components/ResourceList/ResourceListHeader'
-export * from './Components/ResourceList/ResourceListSorter'
-export * from './Components/ResourceList/ResourceCard'
-export * from './Components/ResourceList/ResourceDescription'
-export * from './Components/ResourceList/ResourceName'
-export * from './Components/ResourceList/ResourceEditableName'
+export * from './Components/ResourceList/List/ResourceList'
+export * from './Components/ResourceList/List/ResourceListBody'
+export * from './Components/ResourceList/List/ResourceListHeader'
+export * from './Components/ResourceList/List/ResourceListSorter'
+export * from './Components/ResourceList/Card/ResourceCard'
+export * from './Components/ResourceList/Card/ResourceCardDescription'
+export * from './Components/ResourceList/Card/ResourceCardName'
+export * from './Components/ResourceList/Card/ResourceCardEditableName'
 export * from './Types/index'
